refactor(shared-ui): add ui entrypoint and migrate Stars component
