+++ src/shared/ui/Stars/Stars.jsx
