+++ src/shared/ui/index.js
@@ -1,0 +1,5 @@
+export { default as LoadMoreButton } from './LoadMoreButton/LoadMoreButton';
+export { default as Pagination } from './Pagination/Pagination';
+export { default as PriceSelector } from './PriceSelector/PriceSelector';
+export { default as ProductGallery } from './ProductGallery/ProductGallery';
+export { default as Stars } from './Stars/Stars';
