+++ src/widgets/layout/banner/ui/Banner.jsx
@@ -6,7 +6,7 @@ import Link from 'next/link';
 import { useEffect, useMemo, useRef, useState } from 'react';
 
 import { getBrands } from '@/entities/brand';
-import LoadMoreButton from '@/shared/ui/LoadMoreButton/LoadMoreButton';
+import { LoadMoreButton } from '@/shared/ui';
 
 function SkeletonCard() {
   return (
